+++ public/app/features/alerting/unified/components/amnotifications/AmNotificationsGroup.tsx
@@ -1,11 +1,12 @@
 import { AlertmanagerGroup, AlertState } from 'app/plugins/datasource/alertmanager/types';
-import React, { useState, useMemo } from 'react';
+import React, { useState } from 'react';
 import { GrafanaTheme2 } from '@grafana/data';
 import { useStyles2 } from '@grafana/ui';
 import { css } from '@emotion/css';
 import { AlertLabels } from '../AlertLabels';
 import { AmNotificationsAlertsTable } from './AmNotificationsAlertsTable';
 import { CollapseToggle } from '../CollapseToggle';
+import { AmNotificationsGroupHeader } from './AmNotificationsGroupHeader';
 
 interface Props {
   group: AlertmanagerGroup;
@@ -16,32 +17,6 @@ export const AmNotificationsGroup = ({ alertManagerSourceName, group }: Props) =
   const [isCollapsed, setIsCollapsed] = useState<boolean>(true);
   const styles = useStyles2(getStyles);
 
-  const headerSummary = useMemo(() => {
-    const total = group.alerts.length;
-    const countByStatus = group.alerts.reduce((statusObj, alert) => {
-      if (statusObj[alert.status.state]) {
-        statusObj[alert.status.state] += 1;
-      } else {
-        statusObj[alert.status.state] = 1;
-      }
-      return statusObj;
-    }, {} as Record<AlertState, number>);
-
-    return (
-      <div className={styles.summary}>
-        {`${total} alerts: `}
-        {Object.entries(countByStatus).map(([state, count], index) => {
-          return (
-            <span key={`${group.id}-${index}`} className={styles[state as AlertState]}>
-              {index > 0 && ', '}
-              {`${count} ${state}`}
-            </span>
-          );
-        })}
-      </div>
-    );
-  }, [group, styles]);
-
   return (
     <div className={styles.wrapper}>
       <div className={styles.header}>
@@ -49,7 +24,7 @@ export const AmNotificationsGroup = ({ alertManagerSourceName, group }: Props) =
           <CollapseToggle isCollapsed={isCollapsed} onToggle={() => setIsCollapsed(!isCollapsed)} />
           {Object.keys(group.labels).length ? <AlertLabels labels={group.labels} /> : <span>No grouping</span>}
         </div>
-        {headerSummary}
+        <AmNotificationsGroupHeader group={group} />
       </div>
       {!isCollapsed && (
         <AmNotificationsAlertsTable alertManagerSourceName={alertManagerSourceName} alerts={group.alerts} />
